scripts/requirements.txt
flask==3.0.0
flask-cors==4.0.0
aiohttp==3.9.0
python-dotenv==1.0.0